+++ src/inspect_ai/_view/www/src/samples/SampleDialog.tsx
@@ -1,27 +1,22 @@
 import { ApplicationIcons } from "../appearance/icons";
 import { LargeModal, ModalTool, ModalTools } from "../components/LargeModal";
 
-import { FC, Ref, RefObject, useCallback, useMemo, useRef } from "react";
+import { FC, Ref, useCallback, useEffect, useMemo, useRef } from "react";
 import { ErrorPanel } from "../components/ErrorPanel";
-import { EvalSample } from "../types/log";
+import { useLogSelection, usePrevious, useSampleData } from "../state/hooks";
+import { useStatefulScrollPosition } from "../state/scrolling";
+import { useStore } from "../state/store";
 import { SampleDisplay } from "./SampleDisplay";
-import { SamplesDescriptor } from "./descriptor/samplesDescriptor";
 
 interface SampleDialogProps {
   id: string;
   title: string;
-  sampleStatus: string;
-  sampleError?: Error;
-  sample?: EvalSample;
-  sampleDescriptor: SamplesDescriptor;
   selectedTab?: string;
   setSelectedTab: (tab: string) => void;
   showingSampleDialog: boolean;
   setShowingSampleDialog: (showing: boolean) => void;
   nextSample: () => void;
   prevSample: () => void;
-  sampleScrollPositionRef: RefObject<number>;
-  setSampleScrollPosition: (position: number) => void;
 }
 
 /**
@@ -30,21 +25,55 @@ interface SampleDialogProps {
 export const SampleDialog: FC<SampleDialogProps> = ({
   id,
   title,
-  sample,
-  sampleDescriptor,
   nextSample,
   prevSample,
-  sampleStatus,
-  sampleError,
   showingSampleDialog,
   setShowingSampleDialog,
   selectedTab,
   setSelectedTab,
-  sampleScrollPositionRef,
-  setSampleScrollPosition,
 }) => {
+  // Scroll referernce (attach stateful trackign)
   const scrollRef: Ref<HTMLDivElement> = useRef(null);
+  useStatefulScrollPosition(scrollRef, "sample-dialog");
 
+  // Sample hooks
+  const sampleData = useSampleData();
+  const loadSample = useStore((state) => state.sampleActions.loadSample);
+  const logSelection = useLogSelection();
+
+  // Load sample
+  const prevCompleted = usePrevious(
+    logSelection.sample?.completed !== undefined
+      ? logSelection.sample.completed
+      : true,
+  );
+  const prevLogFile = usePrevious<string | undefined>(logSelection.logFile);
+  useEffect(() => {
+    if (logSelection.logFile && logSelection.sample) {
+      const currentSampleCompleted =
+        logSelection.sample.completed !== undefined
+          ? logSelection.sample.completed
+          : true;
+
+      if (
+        prevLogFile !== logSelection.logFile ||
+        sampleData.sample?.id !== logSelection.sample.id ||
+        sampleData.sample?.epoch !== logSelection.sample.epoch ||
+        currentSampleCompleted !== prevCompleted
+      ) {
+        loadSample(logSelection.logFile, logSelection.sample);
+      }
+    }
+  }, [
+    logSelection.logFile,
+    logSelection.sample?.id,
+    logSelection.sample?.epoch,
+    logSelection.sample?.completed,
+    sampleData.sample?.id,
+    sampleData.sample?.epoch,
+  ]);
+
+  // Tools
   const tools = useMemo<ModalTools>(() => {
     const nextTool: ModalTool = {
       label: "Next Sample",
@@ -100,18 +129,18 @@ export const SampleDialog: FC<SampleDialogProps> = ({
       onkeyup={handleKeyUp}
       visible={showingSampleDialog}
       onHide={onHide}
-      showProgress={sampleStatus === "loading"}
-      initialScrollPositionRef={sampleScrollPositionRef}
-      setInitialScrollPosition={setSampleScrollPosition}
+      showProgress={
+        sampleData.status === "loading" || sampleData.status === "streaming"
+      }
       scrollRef={scrollRef}
     >
-      {sampleError ? (
-        <ErrorPanel title="Sample Error" error={sampleError} />
+      {sampleData.error ? (
+        <ErrorPanel title="Sample Error" error={sampleData.error} />
       ) : (
         <SampleDisplay
           id={id}
-          sample={sample}
-          sampleDescriptor={sampleDescriptor}
+          sample={sampleData.sample}
+          runningEvents={sampleData.running}
           selectedTab={selectedTab}
           setSelectedTab={setSelectedTab}
           scrollRef={scrollRef}