+++ src/inspect_ai/_view/www/src/samples/descriptor/score/ObjectScoreDescriptor.tsx
@@ -52,13 +52,14 @@ export const objectScoreDescriptor = (values: Value2[]): ScoreDescriptor => {
             : String(value);
         scores.push(
           <div
+            key={`score-value-${index}`}
             className={clsx(
               styles.container,
               index + 1 < keys.length ? styles.padded : undefined,
             )}
           >
             <div className={clsx(styles.key, "text-size-smaller")}>{key}</div>
-            <div className={clsx(styles.value, "text-size-title")}>
+            <div className={clsx(styles.value, "text-size-large")}>
               {formattedValue}
             </div>
           </div>,