+++ src/inspect_ai/_view/www/src/samples/scores/SampleScores.tsx
@@ -1,30 +1,26 @@
 import { FC, Fragment } from "react";
 import { SampleSummary } from "../../api/types";
-import { SamplesDescriptor } from "../descriptor/samplesDescriptor";
 
+import { useSampleDescriptor } from "../../state/hooks";
 import styles from "./SampleScores.module.css";
 
 interface SampleScoresProps {
   sample: SampleSummary;
-  sampleDescriptor: SamplesDescriptor;
   scorer: string;
 }
 
-export const SampleScores: FC<SampleScoresProps> = ({
-  sample,
-  sampleDescriptor,
-  scorer,
-}) => {
+export const SampleScores: FC<SampleScoresProps> = ({ sample, scorer }) => {
+  const samplesDescriptor = useSampleDescriptor();
   const scores = scorer
-    ? sampleDescriptor.evalDescriptor
+    ? samplesDescriptor?.evalDescriptor
         .scorerDescriptor(sample, { scorer, name: scorer })
         .scores()
-    : sampleDescriptor.selectedScorerDescriptor(sample).scores();
+    : samplesDescriptor?.selectedScorerDescriptor(sample)?.scores();
 
-  if (scores.length === 1) {
+  if (scores?.length === 1) {
     return scores[0].rendered();
   } else {
-    const rows = scores.map((score) => {
+    const rows = scores?.map((score) => {
       return (
         <Fragment>
           <div style={{ opacity: "0.7" }}>{score.name}</div>